Improving dashboard, prime select and refresh rate, setting ui text

See merge request tuxedocomputers/development/tuxedo-control-center!133

+++ src/ng-app/app/prime-select/prime-select.component.ts
@@ -17,7 +17,7 @@
  * along with TUXEDO Control Center.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import { Component, EventEmitter, OnInit, Output } from "@angular/core";
+import { Component, OnInit } from "@angular/core";
 import { ConfigService } from "../config.service";
 import { UtilsService } from "../utils.service";
 import { TccDBusClientService } from "../tcc-dbus-client.service";
@@ -60,7 +60,7 @@ export class PrimeSelectComponent implements OnInit {
     public async applyGpuProfile(selectedPrimeStatus: string): Promise<void> {
         const status = await this.askProceedDialog();
 
-        if (status !== "APPLY") {
+        if (status === "CANCEL" || status === undefined) {
             this.activeState = this.primeState;
             return;
         }
@@ -80,8 +80,9 @@ export class PrimeSelectComponent implements OnInit {
 
         if (isSuccessful) {
             this.activeState = this.primeState = selectedPrimeStatus;
-            await this.showRebootDialog();
-            this.utils.execCmd("reboot");
+            if (status === "REBOOT") {
+                this.utils.execCmd("reboot");
+            }
         } else {
             this.activeState = this.primeState;
         }
@@ -89,35 +90,24 @@ export class PrimeSelectComponent implements OnInit {
 
     private async askProceedDialog(): Promise<string> {
         const rebootConfig = {
-            title: $localize`Warning`,
-            description: `You will be required to reboot after performing this action. Are you sure you want to apply the selected setting?`,
+            title: $localize`:@@primeSelectAskProceedTitle:Warning`,
+            description: $localize`:@@primeSelectAskProceedDescription:For the change to take effect, your computer must be restarted. If the computer is shut down immediately, unsaved files or changes will be lost. In this case, select 'Reboot later' and shut down your computer at a later time.`,
             labelData: [
                 {
-                    label: `Cancel`,
+                    label: $localize`:@@dialogAbort:Cancel`,
                     value: "CANCEL",
                 },
                 {
-                    label: `Apply`,
-                    value: "APPLY",
+                    label: $localize`:@@dialogInstantReboot:Instant Reboot`,
+                    value: "REBOOT",
+                },
+                {
+                    label: $localize`:@@dialogRebootLater:Reboot later`,
+                    value: "NO_REBOOT",
                 },
             ],
         };
         const returnValue = await this.utils.choiceDialog(rebootConfig);
-        return returnValue.value as string;
-    }
-
-    private async showRebootDialog(): Promise<string> {
-        const rebootConfig = {
-            title: $localize`:@@primeSelectDialogRebootTitle:Completed`,
-            description: `Please note that proceeding with this action will result in a system reboot, which may cause any unsaved work to be lost.`,
-            labelData: [
-                {
-                    label: `Reboot`,
-                    value: "REBOOT",
-                },
-            ],
-        };
-        const returnValue = await this.utils.choiceDialog(rebootConfig, true);
         return returnValue.value as string;
     }
 }
